Fix passing a uint8_t as an int. Breaks on some args.

+++ src/detect.h
@@ -553,7 +553,7 @@ typedef struct DetectTagHostCtx_ {
 /** \brief main detection engine ctx */
 typedef struct DetectEngineCtx_ {
     uint8_t flags;
-    uint8_t failure_fatal;
+    int failure_fatal;
 
     Signature *sig_list;
     uint32_t sig_cnt;
